Extracted table data generation outside Table component
